# web/openapi/faultevolve.openapi.json
{
  "components": {
    "schemas": {
...
      "AppendRoundsRequest": {
        "additionalProperties": false,
        "description": "Ask a run to evolve further (PRD §16.1).\n\n``additional_iterations`` is a *delta*, not a new total. The original\ncompletion summary is never rewritten, so \"this run did N rounds, then M\nmore\" stays auditable after the fact.",
        "properties": {
          "additional_iterations": {
            "exclusiveMinimum": 0,
            "maximum": 500,
            "title": "Additional Iterations",
            "type": "integer"
          }
        },
        "required": [
          "additional_iterations"
        ],
        "title": "AppendRoundsRequest",
        "type": "object"
      },
...
      "ArtifactState": {
        "description": "Availability of one artifact inside a run directory.",
        "enum": [
          "available",
          "missing",
          "invalid"
        ],
        "title": "ArtifactState",
        "type": "string"
      },
      "ContractWarning": {
        "additionalProperties": false,
        "description": "A field the current version could not fill.\n\nThe UI renders these as \"当前版本未记录\" instead of inventing a value.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "title": "ContractWarning",
        "type": "object"
      },
      "ControlAction": {
        "description": "Lifecycle actions the API accepts.\n\n``PAUSE``/``RESUME`` are declared so the API can answer with a stable\n``not_supported`` error instead of pretending (TODO §7.4). The engine has no\nsafe pause point today.",
        "enum": [
          "cancel",
          "pause",
          "resume"
        ],
        "title": "ControlAction",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunBudgetResponse": {
        "additionalProperties": false,
        "description": "Token and time accounting (TODO §5.7). All values from RunSummary.",
        "properties": {
          "avg_smoke_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Avg Smoke Time S"
          },
          "budget_stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Budget Stop Reason"
          },
          "crossover_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Crossover Tokens"
          },
          "discovery_token_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Discovery Token Share"
          },
          "discovery_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Discovery Tokens"
          },
          "elite_archive_sizes": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Elite Archive Sizes",
            "type": "object"
          },
          "expand_skipped_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expand Skipped Budget"
          },
          "generate_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Generate Tokens"
          },
          "knowledge_token_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Knowledge Token Share"
          },
          "knowledge_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Knowledge Tokens"
          },
          "reflect_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reflect Tokens"
          },
          "repair_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Repair Tokens"
          },
          "self_fix_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Self Fix Tokens"
          },
          "smoke_time_s_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Smoke Time S Total"
          },
          "stage_cap_hits": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Stage Cap Hits",
            "type": "object"
          },
          "tokens_per_ros_point": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tokens Per Ros Point"
          },
          "tokens_per_score_gain": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tokens Per Score Gain"
          },
          "total_eval_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Total Eval Time S"
          },
          "total_llm_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Total Llm Latency Ms"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Total Tokens"
          },
          "tournament_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tournament Tokens"
          },
          "wall_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Wall Time S"
          }
        },
        "title": "RunBudgetResponse",
        "type": "object"
      },
...
      "RunControlResponse": {
        "additionalProperties": false,
        "description": "Current lifecycle state after (or instead of) a control action.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ControlAction"
          },
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Status"
          }
        },
        "required": [
          "run_id",
          "action",
          "applied"
        ],
        "title": "RunControlResponse",
        "type": "object"
      },
...
      "RunCreateRequest": {
        "additionalProperties": false,
        "description": "Start one evolution run.\n\n``config_patch`` is a *patch* on top of a server-built preset; the server\nmerges it and re-validates against ``EvolveConfig``.",
        "properties": {
          "config_patch": {
            "additionalProperties": true,
            "title": "Config Patch",
            "type": "object"
          },
          "mock_llm": {
            "default": false,
            "title": "Mock Llm",
            "type": "boolean"
          },
          "preset": {
            "title": "Preset",
            "type": "string"
          },
          "resume_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Resume Run Id"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Seed"
          },
          "server_profile_id": {
            "title": "Server Profile Id",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "server_profile_id",
          "preset"
        ],
        "title": "RunCreateRequest",
        "type": "object"
      },
      "RunCreateResponse": {
        "additionalProperties": false,
        "description": "A run the service accepted and started (TODO 7.3 step 9 -> HTTP 202).\n\nNo filesystem path appears here, on purpose: the UI addresses a run by id\nand asks the server for whatever else it needs. ``execution_backend`` is the\ntransport that actually ran (``local_cloud`` / ``ssh_cloud``), not an echo\nof the request.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "execution_backend": {
            "title": "Execution Backend",
            "type": "string"
          },
          "resumed_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Resumed From"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "execution_backend",
          "task_id",
          "created_at"
        ],
        "title": "RunCreateResponse",
        "type": "object"
      },
      "RunDataState": {
        "description": "Which artifacts back a run directory (TODO §2.4).\n\nKept separate from the run's own status: a run can be ``finished`` and still\nbe ``live``-only if its ``run_summary.json`` was never written.",
        "enum": [
          "complete",
          "live",
          "unreadable"
        ],
        "title": "RunDataState",
        "type": "string"
      },
      "RunDetailResponse": {
        "additionalProperties": false,
        "description": "Everything the run-level pages read out of one run directory.\n\nEach group is ``None`` when the artifact for it is unavailable, so the UI can\ndistinguish \"read but zero\" from \"cannot be read\".",
        "properties": {
          "artifact_states": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ArtifactState"
            },
            "title": "Artifact States",
            "type": "object"
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunBudgetResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "data_state": {
            "$ref": "#/components/schemas/RunDataState"
          },
          "discovery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunDiscoveryResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "judge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunJudgeResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "knowledge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunKnowledgeResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "operators": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOperatorResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOutcomeResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "repair": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRepairResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ContractWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "run_id",
          "data_state"
        ],
        "title": "RunDetailResponse",
        "type": "object"
      },
...
      "RunKnowledgeResponse": {
        "additionalProperties": false,
        "description": "Literature knowledge-package usage.",
        "properties": {
          "card_adoption_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Card Adoption Rate"
          },
          "card_stats": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCardStatResponse"
            },
            "title": "Card Stats",
            "type": "array"
          },
          "cards_adopted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cards Adopted"
          },
          "cards_adopted_valid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cards Adopted Valid"
          },
          "cards_offered": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cards Offered"
          }
        },
        "title": "RunKnowledgeResponse",
        "type": "object"
      },
...
      "RunSummaryResponse": {
        "additionalProperties": false,
        "description": "One row of the run listing (TODO §2.4).",
        "properties": {
          "best_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Best Score"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Created At"
          },
          "data_state": {
            "$ref": "#/components/schemas/RunDataState"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Finished At"
          },
          "improvement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Improvement"
          },
          "iterations_done": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Iterations Done"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Status"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Stop Reason"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Task Id"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ContractWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "data_state"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only control and replay surface for the FaultEvolve engine. Scores are produced by the task evaluator and only ever read back here.",
    "title": "FaultEvolve Web API",
    "version": "1.11.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "Every run directory one level below the configured roots.\n\nAn unreadable run still appears, marked ``unreadable``: hiding it would\nmake a broken run look like a run that never existed.",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryResponse"
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "列出可回放的运行",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "TODO 7.3: validate -> resolve paths -> readiness -> spawn -> 202.\n\n``202`` rather than ``201``: the run has been accepted and a process is\nrunning, but nothing is finished. The body carries the id the UI then\npolls ``/live`` with.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建并启动一次运行",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "description": "Grouped view. Each group is ``null`` when its artifact is unreadable.",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取单个运行的详情",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/cancel": {
      "post": {
        "description": "TODO test #13: stopping a run never deletes what it produced.",
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取消运行（保留已有制品）",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/pause": {
      "post": {
        "description": "Always ``409`` with a reason. The engine has no safe pause point.\n\nDeclared rather than omitted so the UI has a stable answer to render:\na 404 would read as \"wrong URL\", which is not what is happening.",
        "operationId": "pause_run_api_runs__run_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "暂停运行（本版本不支持）",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/resume": {
      "post": {
        "description": "Continue the same experiment under the same id (``--resume``).",
        "operationId": "resume_run_api_runs__run_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "从已有运行恢复",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/rounds": {
      "post": {
        "description": "PRD §16.1. ``additional_iterations`` is a delta, not a new total.",
        "operationId": "append_rounds_api_runs__run_id__rounds_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendRoundsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "追加进化轮次",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}